requirements.txt
aiogram>=3.13
python-dotenv>=1.0.0
aiosqlite>=0.20
apscheduler>=3.10.4